requirements.txt
altair==4.1.0; python_version >= "3.6"
appnope==0.1.2; platform_system == "Darwin" and python_version >= "3.7" and sys_platform == "darwin"
argon2-cffi==20.1.0; python_version >= "3.6"
astor==0.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
async-generator==1.10; python_full_version >= "3.6.1" and python_version >= "3.6"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
backcall==0.2.0; python_version >= "3.7"
base58==2.1.0; python_version >= "3.6"
bleach==3.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
blinker==1.4; python_version >= "3.6"
cachetools==4.2.2; python_version >= "3.6" and python_version < "4.0"
certifi==2021.5.30; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
cffi==1.14.5; python_full_version >= "3.6.1" and python_version >= "3.6" and implementation_name == "pypy"
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
click==7.1.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0"
cycler==0.10.0; python_version >= "3.7"
decorator==5.0.9; python_version >= "3.7"
defusedxml==0.7.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
entrypoints==0.3; python_version >= "3.6"
gitdb==4.0.7; python_version >= "3.6"
gitpython==3.1.17; python_version >= "3.6"
idna==2.10; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
imageio==2.9.0; python_version >= "3.7"
ipykernel==5.5.5; python_version >= "3.6"
ipython-genutils==0.2.0; python_version >= "3.7"
ipython==7.24.1; python_version >= "3.7"
ipywidgets==7.6.3; python_version >= "3.6"
jedi==0.18.0; python_version >= "3.7"
jinja2==3.0.1; python_version >= "3.6"
jsonschema==3.2.0; python_version >= "3.6"
jupyter-client==6.2.0; python_full_version >= "3.6.1" and python_version >= "3.6"
jupyter-core==4.7.1; python_full_version >= "3.6.1" and python_version >= "3.6"
jupyterlab-pygments==0.1.2; python_version >= "3.6"
jupyterlab-widgets==1.0.0; python_version >= "3.6"
kiwisolver==1.3.1; python_version >= "3.7"
markupsafe==2.0.1; python_version >= "3.6"
matplotlib-inline==0.1.2; python_version >= "3.7"
matplotlib==3.4.2; python_version >= "3.7"
mistune==0.8.4; python_version >= "3.6"
nbclient==0.5.3; python_full_version >= "3.6.1" and python_version >= "3.6"
nbconvert==6.0.7; python_version >= "3.6"
nbformat==5.1.3; python_full_version >= "3.6.1" and python_version >= "3.6"
nest-asyncio==1.5.1; python_full_version >= "3.6.1" and python_version >= "3.6"
networkx==2.5; python_version >= "3.7"
notebook==6.4.0; python_version >= "3.6"
numpy==1.20.3; python_full_version >= "3.7.1" and python_version >= "3.7"
opencv-python==4.5.2.54; python_version >= "3.6"
packaging==20.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pandas==1.2.4; python_full_version >= "3.7.1" and python_version >= "3.6"
pandocfilters==1.4.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
parso==0.8.2; python_version >= "3.7"
patsy==0.5.1; python_version >= "3.6"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.7"
pickleshare==0.7.5; python_version >= "3.7"
pillow==8.2.0; python_version >= "3.7"
pip-chill==1.0.1
plotly-express==0.4.1
plotly==4.14.3
prometheus-client==0.11.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
prompt-toolkit==3.0.18; python_full_version >= "3.6.1" and python_version >= "3.7"
protobuf==3.17.3; python_version >= "3.6"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.7" and os_name != "nt"
py==1.10.0; python_full_version >= "3.6.1" and python_version >= "3.6" and implementation_name == "pypy"
pycparser==2.20; python_full_version >= "3.6.1" and python_version >= "3.6" and implementation_name == "pypy"
pydeck==0.6.2; python_version >= "3.6"
pygments==2.9.0; python_version >= "3.7"
pyparsing==2.4.7; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
pyrsistent==0.17.3; python_version >= "3.6"
python-dateutil==2.8.1; python_full_version >= "3.7.1" and python_version >= "3.6" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7")
pytz==2021.1; python_full_version >= "3.7.1" and python_version >= "3.6"
pywavelets==1.1.1; python_version >= "3.7"
pywin32==301; python_full_version >= "3.6.1" and python_version >= "3.6" and sys_platform == "win32"
pywinpty==1.1.2; os_name == "nt" and python_version >= "3.6"
pyzmq==22.1.0; python_full_version >= "3.6.1" and python_version >= "3.6"
requests==2.25.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
retrying==1.3.3
scikit-image==0.18.2; python_version >= "3.7"
scipy==1.6.1; python_version >= "3.7"
send2trash==1.5.0; python_version >= "3.6"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
smmap==4.0.0; python_version >= "3.6"
statsmodels==0.12.2; python_version >= "3.6"
streamlit==0.82.0; python_version >= "3.6"
terminado==0.10.1; python_version >= "3.6"
testpath==0.5.0; python_version >= "3.6"
tifffile==2021.6.14; python_version >= "3.7"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
toolz==0.11.1; python_version >= "3.6"
torch==1.9.0; python_full_version >= "3.6.2"
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1; python_full_version >= "3.6.1" and python_version >= "3.6"
traitlets==5.0.5; python_full_version >= "3.6.1" and python_version >= "3.7"
typing-extensions==3.10.0.0; python_full_version >= "3.6.2"
tzlocal==2.1; python_version >= "3.6"
urllib3==1.26.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
validators==0.18.2; python_version >= "3.6"
watchdog==2.1.2; platform_system != "Darwin" and python_version >= "3.6"
wcwidth==0.2.5; python_full_version >= "3.6.1" and python_version >= "3.7"
webencodings==0.5.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
widgetsnbextension==3.5.1; python_version >= "3.6"
